review 수정, 가게 별점, 예약 관련 기능 추가

+++ src/main/java/com/nadri/restaurant/vo/Restaurant.java
@@ -28,7 +28,7 @@ public class Restaurant {
 	private String tel;			// 전화번호
 	private String openTime;	// 영업시간
 	private int sales;			// 매출양
-	private int starPoint;		// 별점
+	private double starPoint;		// 별점
 	private double lat;			// 위도 mapx
 	private double lon;			// 경도 mapy
 	private String categoryName;
